Change namespacing of Tags::CollectRemoteInventoryResources and refactor
New namespacing: Tags::Topology::RemoteInventory
