+++ spec/services/tags/topology/remote_inventory_spec.rb
@@ -1,4 +1,4 @@
-describe Tags::CollectRemoteInventoryResources, :type => :service do
+describe Tags::Topology::RemoteInventory, :type => :service do
   let(:subject) { described_class.new(task) }
 
   before do
